move burndown chart
